Step 20: Remove optional from price variable and add null object
This helps to clean up the implementation logic of having to check for
null however due to the nature of the switch statement this means that a
null object pattern has to be introduced. This is simple since
inheritance is being used but may change in the future when we look at
using a protocol instead.